ENH:add slack bot CRUD API, move event listner as controller
